(Almost) Complete Web App
Added the HTML pages, style.css file, made the raw app.py code. The code includes OAuth for Spotify and Youtube APIs. The app can get the playlists from Spotify, create a playlist in youtube and add the videos of the musics to that playlist.
Obs.: Still need to improve the code to handle the quota limits and the time to convert.